JVM_IR: Initial support of crossinline suspend lambdas
They are only supported if captured by suspend lambdas.
Generate IrClasses for suspend lambdas inside their call-site functions,
otherwise, they will not be generated before transformation and the
inliner will fail to find them.
Because of that, lowerings which transform call-site function are likely
to change reference to lowered suspend lambdas. Hence, do not rely on
references to lowered suspend lambdas, instead, rely on names.
Do not generate secondary constructor for suspend lambdas, otherwise,
the inliner is unable to transform them.
Refactor CoroutineTransformerMethodVisitor, so it no longer depends on
PSI.