Set up artifact transforms for incremental Kotlin compile
THIS IS A PROTOTYPE for high-level review. Details will be polished
later.

While there are still many TODOs, this prototype is FULLY FUNCTIONAL for
a simple but key scenario: changing a method signature in a Kotlin or
Java source file in a subproject.

There are 5 key steps (please search for "INCR_KOTLIN_COMPILE_BOOKMARK"
in the code):
  - Step 1 [Gradle side]: Create classpath snapshots, using artifact
    transforms
  - Step 2 [Gradle side]: Compute classpath snapshot changes
  - Step 3 [Gradle side]: Pass classpath snapshot changes to Kotlin
    compile daemon
  - Step 4 [Kotlin daemon]: Retrieve classpath snapshot changes from the
    Gradle side
  - Step 5 [Kotlin daemon]: Use classpath snapshot changes for
    incremental Kotlin compile

Bug: [To create]
Test: [To create]